requirements.txt
playwright==1.53.0
PyYAML==6.0.2
yagmail==0.15.293
schedule==1.2.2
requests==2.32.4 